backend/requirements.txt
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
autopep8==2.3.1
bcrypt==4.2.0
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachelib==0.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
checksumdir==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Excel==0.0.7
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Mailman==1.1.1
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
Flask-WTF==1.2.1
fonttools==4.54.1
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
lml==0.1.0
Mako==1.3.5
MarkupSafe==2.1.5
mkdocs-material-extensions==1.3.1
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pip-review==1.3.0
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pycparser==2.22
pydyf==0.11.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.9.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
texttable==1.7.0
tinycss2==1.3.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.44.0
WTForms==3.1.2
zope.event==5.0
zope.interface==7.0.3
zopfli==0.2.3
